fix: can't transfer configProvider in the content of Modal/Message/Notification
