+++ components/Modal/config.ts
@@ -1,17 +1,3 @@
-import type { ConfigProviderProps } from '../ConfigProvider';
-
-let configProvider = {} as ConfigProviderProps;
-
-export function setConfigProviderProps(configProviderProps: ConfigProviderProps) {
-  configProvider = {
-    ...configProviderProps,
-  };
-}
-
-export function getConfigProviderProps() {
-  return configProvider;
-}
-
 export type ModalConfigType = {
   prefixCls?: string;
   simple?: boolean;
