+++ components/_util/transferConfigProvider.ts
@@ -1,0 +1,13 @@
+import type { ConfigProviderProps } from '../ConfigProvider';
+
+let configProvider = {} as ConfigProviderProps;
+
+export function setConfigProviderProps(configProviderProps: ConfigProviderProps) {
+  configProvider = {
+    ...configProviderProps,
+  };
+}
+
+export function getConfigProviderProps() {
+  return configProvider;
+}
